migrate to a rule system
this allows implementing simple validation rules entirely in configuration without the need to provide a webhook receiver. yet, configuring a webhook is not significantly more difficult than before.

I considered using a scripting language or even wasm to allow arbitrary programs for validation, but I deemed setting up the necessary isolation from the server and other hook executions to complex to be worth it. Use a webhook if you need this.

see #14